VS Code: Work around inlay hint display bug
If VS Code opens a document with no inlay hints visible, it won't
request new inlay hints until the code window is redrawn. This happens
even if you issue `workspace/inlayHint/refresh`

To work around this, we set a zero-width space hint in the document when
no hints would otherwise be shown. This behavior is VS Code only.

This changeset also makes the `prefabPromise` only be settled after the
`onUpdate` fires. This should prevent some race conditions down the
line.